added collector class

+++ src/main/resources/XLRProfile/__init__.py
@@ -11,9 +11,114 @@ import requests
 
 import com.xhaus.jyson.JysonCodec as json
 from requests.auth import HTTPBasicAuth
+
 import com.xebialabs.xlrelease.api.XLReleaseServiceHolder as XLReleaseServiceHolder
 import com.xebialabs.deployit.repository.SearchParameters as SearchParameters
 import com.xebialabs.deployit.plugin.api.reflect.Type as Type
+
+class CollectorError(Exception):
+    """Exception raised for errors in the input.
+
+    Attributes:
+        expr -- input expression in which the error occurred
+        msg  -- explanation of the error
+    """
+
+    def __init__(self,value, msg):
+        self.value = value
+        self.msg = msg
+
+    def __str__(self):
+        return "%s : %s" % (self.msg, self.value)
+
+class Collector():
+
+    attributes = []
+    optional_attributes = []
+
+    def __init__(self, **params):
+       self.validate(**params)
+
+    def validate(self, **params):
+        for a in self.attributes:
+            if params.has_key(a) is not True:
+                raise CollectorError(a, "unable to satisfy all attributes")
+
+    def resolve(self):
+        return None
+
+
+class JsonCollector(Collector):
+
+    attributes = ["url", "path"]
+    optional_attributes = []
+
+    def __init__(self, **params):
+
+        Collector.__init__(self, **params)
+
+        self.requests_params = {}
+
+        if params.has_key('username') and  params.has_key('password'):
+            self.requests_params['auth'] = HTTPBasicAuth(params['username'], params['password'])
+
+        self.__collector_params = params
+
+
+    def resolve(self):
+        """
+        resolve the collector
+        :return: string
+        """
+        return self.get_path(self.load_json(self.__collector_params['url']), self.__collector_params['path'])
+
+
+    def load_json(self, url):
+        try:
+            response = requests.get(url, verify=False, **self.requests_params)
+
+            response.raise_for_status()
+        except requests.exceptions.HTTPError as e:
+            print "unable to retrieve json from url: %s" % url
+            print e.request
+            return None
+        try:
+            print "%s responded with:" % url
+            print response.text
+            return json.loads(str(response.text))
+        except Exception:
+            print "unable to decode information provided by %s" % url
+            return None
+        except JSONDecodeError:
+            print "error"
+            return None
+
+    def get_path(self, json, path):
+
+        if (type(path) is str) or (type(path) is unicode):
+            path = str(path).split('/')
+
+        field = path.pop(0)
+
+        try:
+            if json.has_key(field):
+                if type(json[field]) == dict:
+                    return self.get_path(json[field], path)
+
+                elif type(json[field]) == list:
+                    if len(json[field]) < 2 :
+                        print "found %s" % (json[field][0])
+                        return str(json[field][0])
+
+                elif len(path) == 0:
+                    print "found %s using path %s" % (field, path)
+                    return str(json[field])
+            else:
+                print "the requested path of %s could not be found in the json document. returning None instead"
+                return None
+        except Exception:
+            print "Error encountered during resolution"
+
 
 
 
@@ -50,7 +155,10 @@ class XLRProfile(collections.MutableMapping):
 
 
     def __getitem__(self, key):
-        return self.store[self.__keytransform__(key)]
+        if self.store.has_key(self.__keytransform__(key)):
+            return self.store[self.__keytransform__(key)]
+        else:
+            return None
 
     def __setitem__(self, key, value):
         self.store[self.__keytransform__(key)] = value
@@ -74,6 +182,7 @@ class XLRProfile(collections.MutableMapping):
         self.store['variables'][key] = value
 
     def toggles(self):
+
         return self.__getitem__('toggles')
 
     def load_from_url(self, url):
@@ -82,7 +191,6 @@ class XLRProfile(collections.MutableMapping):
         :param url:
         :return: dict: profile
         """
-        print 1
         response = requests.get(url, verify=False)
         response.raise_for_status()
 
@@ -107,6 +215,10 @@ class XLRProfile(collections.MutableMapping):
         for key, val in self.variables():
             if type(val) == dict:
                 solution = self.resolve_variable(val)
+                if solution == None:
+                    print "value for %s could not be found using the specified collector" % key
+                else:
+                    print "retrieved value: %s for %s" % (solution, key)
                 self.set_variable(key, solution)
 
     def resolve_variable(self, **params):
@@ -130,94 +242,28 @@ class XLRProfile(collections.MutableMapping):
         # act on the collector parameter
         if params.has_key('collector'):
             # if there is no type specified just pass it to the json collector, which is our default choice
+            # if not params['collector'].has_key('type'):
+            #     collector_val = self.handle_json_collector(**params['collector'])
+            # if params['collector'].has_key('type'):
+            #     collector_val = self.handle_json_collector(**params['collector'])
+            # else:
+            #     print "collector type is not supported.... yet!!!"
+            #     sys.exit(2)
+            # if there is no type specified just pass it to the json collector, which is our default choice
             if not params['collector'].has_key('type'):
-                collector_val = self.handle_json_collector(**params['collector'])
+                col_params = params['collector']
+                collector_val = JsonCollector(**col_params).resolve()
             if params['collector'].has_key('type'):
-                collector_val = self.handle_json_collector(**params['collector'])
+                col_params = params['collector']
+                collector_val = JsonCollector(**col_params).resolve()
             else:
                 print "collector type is not supported.... yet!!!"
-                sys.exit(2)
+                pass
 
         if collector_val:
             ret_val = collector_val
-
         return ret_val
 
-    def handle_json_collector(self, **params):
-        """
-        retrieve a template variable by means of json rest call
-        :username: username if we need to authenticate (optional)
-        :password: password if we need to authenticate (optional)
-        :url: url to hit when retrieving the json
-        :path: path to the field we want to retrieve in the json file
-        :param params:
-        :return:
-        """
-        requests_params = {}
-
-        # if we need authentication then lets set it up
-        if params.has_key('username') and params.has_key('password'):
-            requests.params['auth'] = HTTPBasicAuth(params['username'], params['password'])
-
-        if params.has_key('url'):
-            response = requests.get(params['url'], verify=False, **requests_params)
-
-            response.raise_for_status()
-
-            try:
-                inputDictionary = json.loads(str(response.text))
-            except Exception:
-                print "unable to decode information provided by %s" % params['url']
-                return None
-
-            # if the path parameter was given parse it and see if we can retrieve a value . if path was not set return None
-            # i could have chosen to return the json in its raw form, but that might lead to unwanted situations and code injection
-            if params.has_key('path'):
-
-                # get a list with the fields where looking for
-                field_list = params['path'].split('/')
-                len_field = len(field_list)
-                field_counter = 1
-
-                # loop over the field list and see if we can go down the rabbit hole to retrieve the value where looking for
-                for field in field_list:
-                    if inputDictionary.has_key(field):
-
-                        # if the value of the inputDictionary is another dictionary: input that second dict into inputDictionary
-                        # and take it back into the loop
-                        if type(inputDictionary[field]) == dict:
-                            inputDictionary = inputDictionary[field]
-
-                        # if the value is a loop .. check if it only contains one value .. if so return it as a string. If not return None
-                        elif type(inputDictionary[field]) == list:
-
-                            if len(inputDictionary[field]) < 2:
-                                print "found %s using path %s" % (inputDictionary[field], params['path'])
-                                return str(inputDictionary[field][0])
-                            else:
-                                print "found mutiple values:%s using path %s" % (inputDictionary[field], params['path'])
-                                print "this can lead to unpredictable results so no value will be used"
-                                return None
-
-                        # if the field_counter matches the length of the field list we used as input to this loop then we must have found our value
-                        # remember we allready tackled lists and dictionaries so no need to worry about those
-                        elif len_field == field_counter:
-                            print "found %s using path %s" % (inputDictionary[field], params['path'])
-                            return str(inputDictionary[field])
-
-                        else:
-                            print "the requested path of %s could not be found in the json document. returning None instead"
-                            return None
-
-                    else:
-                        print "the requested path of %s could not be found in the json document. returning None instead"
-                        return None
-
-                    field_counter += 1
-
-            else:
-                print "parameter path should be set when using json collector. This parameter seems to be missing so returning None"
-                return None
 
 
     def persist_variables_to_release(self, releaseId):
@@ -239,14 +285,17 @@ class XLRProfile(collections.MutableMapping):
     def handle_toggles(self, releaseId):
 
         release = self.__releaseApi.getRelease(releaseId)
-        for t in self.toggles():
-            task = self.get_task_by_phase_and_title(str(t["phase"]), str(t["task"]), release)
-            if task:
-                print "removing task %s " % task
-                self.__repositoryService.delete(str(task))
+        if self.toggles():
+            for t in self.toggles():
+                task = self.get_task_by_phase_and_title(str(t["phase"]), str(t["task"]), release)
+                if task:
+                    print "removing task %s " % task
+                    self.__repositoryService.delete(str(task))
 
-            else:
-                print "task not found"
+                else:
+                    print "task not found"
+        else:
+            print "no toggles found"
 
 
 
@@ -282,3 +331,90 @@ class XLRProfile(collections.MutableMapping):
                 return tasks[taskTitle]
 
         return False
+
+ # def handle_json_collector(self, **params):
+ #    """
+ #    retrieve a template variable by means of json rest call
+ #    :username: username if we need to authenticate (optional)
+ #    :password: password if we need to authenticate (optional)
+ #    :url: url to hit when retrieving the json
+ #    :path: path to the field we want to retrieve in the json file
+ #    :param params:
+ #    :return:
+ #    """
+ #    requests_params = {}
+ #
+ #
+ #    # if we need authentication then lets set it up
+ #    try:
+ #        if params.has_key('username') and params.has_key('password'):
+ #            requests.params['auth'] = HTTPBasicAuth(params['username'], params['password'])
+ #
+ #        if params.has_key('url'):
+ #            response = requests.get(params['url'], verify=False, **requests_params)
+ #
+ #            response.raise_for_status()
+ #
+ #            try:
+ #                print "%s responded with:" % params['url']
+ #                print response.text
+ #                inputDictionary = json.loads(str(response.text))
+ #            except Exception:
+ #                print "unable to decode information provided by %s" % params['url']
+ #                return None
+ #            except JSONDecodeError:
+ #                print "error"
+ #                return None
+ #
+ #            # if the path parameter was given parse it and see if we can retrieve a value . if path was not set return None
+ #            # i could have chosen to return the json in its raw form, but that might lead to unwanted situations and code injection
+ #            if params.has_key('path'):
+ #
+ #                # get a list with the fields where looking for
+ #                field_list = params['path'].split('/')
+ #                len_field = len(field_list)
+ #                field_counter = 1
+ #
+ #                # loop over the field list and see if we can go down the rabbit hole to retrieve the value where looking for
+ #                for field in field_list:
+ #                    if inputDictionary.has_key(field):
+ #
+ #                        # if the value of the inputDictionary is another dictionary: input that second dict into inputDictionary
+ #                        # and take it back into the loop
+ #                        if type(inputDictionary[field]) == dict:
+ #                            inputDictionary = inputDictionary[field]
+ #
+ #                        # if the value is a loop .. check if it only contains one value .. if so return it as a string. If not return None
+ #                        elif type(inputDictionary[field]) == list:
+ #
+ #                            if len(inputDictionary[field]) < 2:
+ #                                print "found %s using path %s" % (inputDictionary[field], params['path'])
+ #                                return str(inputDictionary[field][0])
+ #                            else:
+ #                                print "found mutiple values:%s using path %s" % (inputDictionary[field], params['path'])
+ #                                print "this can lead to unpredictable results so no value will be used"
+ #                                return None
+ #
+ #                        # if the field_counter matches the length of the field list we used as input to this loop then we must have found our value
+ #                        # remember we allready tackled lists and dictionaries so no need to worry about those
+ #                        elif len_field == field_counter:
+ #                            print "found %s using path %s" % (inputDictionary[field], params['path'])
+ #                            return str(inputDictionary[field])
+ #
+ #                        else:
+ #                            print "the requested path of %s could not be found in the json document. returning None instead"
+ #                            return None
+ #
+ #                    else:
+ #                        print "the requested path of %s could not be found in the json document. returning None instead"
+ #                        return None
+ #
+ #                    field_counter += 1
+ # d
+ #            else:
+ #                print "parameter path should be set when using json collector. This parameter seems to be missing so returning None"
+ #                return None
+ #
+ #    except Exception:
+ #        print "Error encounterd trying to handle json collector for %s" % params['url']
+ #        return None
